Add ctor, method and property tests to AsciiEncoding
- Ctor
- EncoderFallback, DecoderFallback
- WebName, CodePage, EncodingName, IsSingleByte
- Clone, Equals, GetPreamble
