+++ src/System.Text.Encoding/tests/System.Text.Encoding.Tests.csproj
@@ -17,15 +17,13 @@
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="ASCIIEncoding\ASCIIEncodingProperties.cs" />
-    <Compile Include="ASCIIEncoding\ASCIIEncodingEquals.cs" />
     <Compile Include="ASCIIEncoding\ASCIIEncodingEncode.cs" />
     <Compile Include="ASCIIEncoding\ASCIIEncodingDecode.cs" />
     <Compile Include="ASCIIEncoding\ASCIIEncodingGetDecoder.cs" />
-    <Compile Include="ASCIIEncoding\ASCIIEncodingGetPreamble.cs" />
     <Compile Include="ASCIIEncoding\ASCIIEncodingGetEncoder.cs" />
     <Compile Include="ASCIIEncoding\ASCIIEncodingGetMaxByteCount.cs" />
     <Compile Include="ASCIIEncoding\ASCIIEncodingGetMaxCharCount.cs" />
+    <Compile Include="ASCIIEncoding\ASCIIEncodingTests.cs" />
     <Compile Include="Decoder\DecoderConvert2.cs" />
     <Compile Include="Decoder\DecoderCtor.cs" />
     <Compile Include="Decoder\DecoderGetCharCount2.cs" />
